Move isCustomDs and isRuntimeDs functions to common
