+++ packages/node/src/indexer/fetch.service.ts
@@ -8,6 +8,8 @@ import { ApiPromise } from '@polkadot/api';
 import {
   isRuntimeDataSourceV0_2_0,
   RuntimeDataSourceV0_0_1,
+  isCustomDs,
+  isRuntimeDs,
 } from '@subql/common';
 import {
   SubqlCallFilter,
@@ -22,12 +24,7 @@ import { NodeConfig } from '../configure/NodeConfig';
 import { SubqueryProject } from '../configure/project.model';
 import { getLogger } from '../utils/logger';
 import { profiler, profilerWrap } from '../utils/profiler';
-import {
-  isBaseHandler,
-  isCustomDs,
-  isCustomHandler,
-  isRuntimeDs,
-} from '../utils/project';
+import { isBaseHandler, isCustomHandler } from '../utils/project';
 import { delay } from '../utils/promise';
 import * as SubstrateUtil from '../utils/substrate';
 import { getYargsOption } from '../yargs';
